+++ hemi/website/test/test.go
@@ -1,9 +1,0 @@
-package test
-
-import (
-	"fmt"
-)
-
-func Main() {
-	fmt.Println("tests")
-}